code optimisation changes reverted for body component.
